yeonsup/fix : 버그 수정 1. 당첨금액이 Int 타입 범위 초과로 long으로 변경 2. 담청 개수 매개변수 잘못 넘김으로인한 버그 수정

+++ src/main/java/lotto/model/WiningType.java
@@ -16,9 +16,9 @@ public enum WiningType {
     private String name;
     private int correctCount;
     private boolean hasBonus;
-    private int amount;
+    private long amount;
 
-    WiningType(String name, int correctCount, boolean hasBonus, int amount) {
+    WiningType(String name, int correctCount, boolean hasBonus, long amount) {
         this.name = name;
         this.correctCount = correctCount;
         this.hasBonus = hasBonus;
@@ -52,7 +52,7 @@ public enum WiningType {
         return correctCount;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
 }
